+++ src/main/java/com/example/proyectodbp/station/domain/StationService.java
@@ -3,75 +3,47 @@ package com.example.proyectodbp.station.domain;
 import com.example.proyectodbp.events.HelloEmailEvent;
 import com.example.proyectodbp.exceptions.ResourceNotFoundException;
 import com.example.proyectodbp.exceptions.UnauthorizedOperationException;
+import com.example.proyectodbp.exceptions.UniqueResourceAlreadyExist;
 import com.example.proyectodbp.route.domain.Route;
 import com.example.proyectodbp.route.infraestructure.RouteRepository;
 import com.example.proyectodbp.station.dto.NewStationRequestDto;
 import com.example.proyectodbp.station.dto.StationDto;
 import com.example.proyectodbp.station.infraestructure.StationRepository;
-import com.example.proyectodbp.user.domain.Role;
-import com.example.proyectodbp.user.domain.User;
-import com.example.proyectodbp.user.infraestructure.UserRepository;
 import com.example.proyectodbp.utils.AuthorizationUtils;
 import org.modelmapper.ModelMapper;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class StationService {
     private final StationRepository stationRepository;
     private final RouteRepository routeRepository;
-    private final UserRepository<User> userRepository;
     private final AuthorizationUtils authorizationUtils;
-    private final ModelMapper modelMapper;
-
+    private final ModelMapper modelMapper = new ModelMapper();
     private final ApplicationEventPublisher applicationEventPublisher;
 
-    public StationService(StationRepository stationRepository, RouteRepository routeRepository, UserRepository<User> userRepository, AuthorizationUtils authorizationUtils, ModelMapper modelMapper, ApplicationEventPublisher applicationEventPublisher) {
+    public StationService(StationRepository stationRepository, RouteRepository routeRepository, AuthorizationUtils authorizationUtils, ApplicationEventPublisher applicationEventPublisher) {
         this.stationRepository = stationRepository;
         this.routeRepository = routeRepository;
-        this.userRepository = userRepository;
         this.authorizationUtils = authorizationUtils;
-        this.modelMapper = modelMapper;
         this.applicationEventPublisher = applicationEventPublisher;
     }
 
-    public String createStation(NewStationRequestDto stationDto) {
-        String username = authorizationUtils.getCurrentUserEmail();
-        if(username == null) {
-            throw new UnauthorizedOperationException("Anonymous User not allowed to access");
+    public String createStation(NewStationRequestDto newStation) {
+        if (stationRepository.findByName(newStation.getName()).isPresent()) {
+            throw new UniqueResourceAlreadyExist("This station already exists");
         }
 
-        // Verifica que el usuario actual sea un DRIVER
-        User user = userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        if(user.getRole() != Role.DRIVER) {
-            throw new UnauthorizedOperationException("No estas autorizado para acceder a este recurso");
-        }
-  
-        if (stationRepository.findByName(stationDto.getName()).isPresent()) {
-            throw new ResourceNotFoundException("This station already exists");
-        }
-
-        Station station = modelMapper.map(stationDto, Station.class);
+        Station station = modelMapper.map(newStation, Station.class);
         stationRepository.save(station);
         String message = "Station created";
-        applicationEventPublisher.publishEvent(new HelloEmailEvent(user.getEmail(), message));
+        applicationEventPublisher.publishEvent(new HelloEmailEvent(authorizationUtils.getCurrentUserEmail(), message));
         return "/station/" + station.getId();
     }
 
     public StationDto getStationInfo(Long id) {
-        String username = authorizationUtils.getCurrentUserEmail();
-        if(username == null) {
-            throw new UnauthorizedOperationException("Anonymous User not allowed to access");
-        }
-
-        // Verifica que el usuario actual sea un DRIVER
-        User user = userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        if(user.getRole() != Role.DRIVER) {
-            throw new UnauthorizedOperationException("No estas autorizado para acceder a este recurso");
-        }
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
 
         Station station = stationRepository
                 .findById(id)
@@ -81,47 +53,31 @@ public class StationService {
     }
 
     public void deleteStation(Long id) {
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
+
         stationRepository.deleteById(id);
     }
 
     public void updateStation(Long id, StationDto stationDto) {
-        // Check if the current user is an admin or the owner of the resource
-        String username = authorizationUtils.getCurrentUserEmail();
-        if(username == null) {
-            throw new UnauthorizedOperationException("Anonymous User not allowed to access");
-        }
-
-        // Verifica que el usuario actual sea un DRIVER
-        User user = userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        if(user.getRole() != Role.DRIVER) {
-            throw new UnauthorizedOperationException("No estas autorizado para acceder a este recurso");
-        }
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
 
         Station stationToUpdate = stationRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("This station does not exist"));
+
         stationToUpdate.setName(stationDto.getName());
         stationToUpdate.setRoutes(stationDto.getRoutes());
         stationRepository.save(stationToUpdate);
         String message = "Station updated";
-        applicationEventPublisher.publishEvent(new HelloEmailEvent(user.getEmail(), message));
+        applicationEventPublisher.publishEvent(new HelloEmailEvent(authorizationUtils.getCurrentUserEmail(), message));
     }
 
     public void addRoute(Long id, String routeName) {
-        // Check if the current user is an admin or the owner of the resource
-        String username = authorizationUtils.getCurrentUserEmail();
-        if(username == null) {
-            throw new UnauthorizedOperationException("Anonymous User not allowed to access");
-        }
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
 
-        // Verifica que el usuario actual sea un DRIVER
-        User user = userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        if(user.getRole() != Role.DRIVER) {
-            throw new UnauthorizedOperationException("No estas autorizado para acceder a este recurso");
-        }
-  
         Station station = stationRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("This station does not exist"));
@@ -133,6 +89,6 @@ public class StationService {
         stationRepository.save(station);
         routeRepository.save(route);
         String message = "Route added to station";
-        applicationEventPublisher.publishEvent(new HelloEmailEvent(user.getEmail(), message));
+        applicationEventPublisher.publishEvent(new HelloEmailEvent(authorizationUtils.getCurrentUserEmail(), message));
     }
 }
